legalai/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
openai
asyncpg
beautifulsoup4
pydantic
bcrypt
python-jose[cryptography]